Tried to fix lack of printing day number in planet meridian passage tables
Currently, the day number (1-31) does not print in the three individual planet tables that show meridian passage, below the Stars table on the left-hand pages. This was an attempt to fix that.

Originally, the date was 'datex[0:8]', which seemed to leave off the two slashes. I change this to 'datex[0:10]' which worked fine for all dates with double digit months. However, January through October (1 through 9) with single digit days appeared with a space and an extra zero.

Next attempt will be to cut the datex string at the first space.